Oxidrone skeleton
One of the priorities of oxidrone is to make it easy for devs to extend
the firmware with more drivers and other components. Therefore the
choice is made to split the project into 3 crates. This will make it
easier in the future to develop and test parts in isolation.

Justfile is used as command line runner. It adds an extra build
dependency, but i expect the investment is worth the lower build
complexity

Matek board will be used as development target for now. Widely available
at a good price + documentation is available.